requirements.txt
pytest
pytest-selenium
testfixtures
